backend/requirements.txt
feedparser==5.2.1
peewee==2.8.0
PyMySQL==0.7.2
python-dateutil==2.5.0
pytz==2015.7
six==1.10.0
wheel==0.24.0
